requirements.txt
streamlit
pandas
numpy
scikit-learn
paho-mqtt